move editor packets back into server
